feat: add global max width to page 404
