+++ src/pages/404/styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+import { globalMaxWidth } from '../../globalStyles';
+
+export const Container = styled.div`
+  ${globalMaxWidth}
+`;
